moved inner classes of TokenStream into separate files
